# Conflicts:
#	src/main/java/com/sorisonsoon/payment/controller/PaymentController.java
#	src/main/java/com/sorisonsoon/payment/domain/entity/Payment.java
#	src/main/java/com/sorisonsoon/payment/domain/repository/PaymentRepository.java
#	src/main/java/com/sorisonsoon/payment/dto/PaymentDTO.java
#	src/main/java/com/sorisonsoon/payment/service/PaymentService.java

+++ src/main/java/com/sorisonsoon/payment/dto/PaymentDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Setter
 public class PaymentDTO {
     private Long paymentId;
-    private int payerId;
+    private Long payerId;
     private int amount;
     private LocalDateTime payedAt;
 }
